+++ doc/release_notes.rst
@@ -4,9 +4,9 @@ Release Notes
 Upcoming Release
 ----------------
 
-
 * The memory-efficiency of the IO to LP/MPS file was further improved. In particular, the function `to_dataframe` is now avoiding unnecessary data copies.
 * The printout of time stamps was modified to be more readable, leaving out the display of seconds and below if not necessary.
+* New solver interface for `SCIP <https://www.scipopt.org/>`. This solver is now supported by `linopy` and can be used with the `solve` function if the `pyscipopt` package is installed. The solver is available for free for general use. See the `SCIP website <https://www.scipopt.org/>` for more information.
 
 
 Version 0.2.5